requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
requests==2.33.1
pillow==12.2.0
pydantic==2.13.3
python-multipart==0.0.26
httpx==0.28.1
python-dotenv==1.2.2
google-genai==1.73.1
duckduckgo_search==8.1.1
mcp==1.27.0
